+++ xgcm/__init__.py
@@ -3,5 +3,5 @@ try:
 except ImportError:
     __version__ = "unknown"
 
-from .grid import Axis, Grid
+from .grid import Grid
 from .grid_ufunc import apply_as_grid_ufunc, as_grid_ufunc